feat(notification): 新增通知模块功能
- 实现通知模块的分页查询、详情查看、创建和删除功能
- 优化通知相关的数据结构和接口定义
-调整通知发送逻辑，支持批量发送给不同用户类型
- 修复通知状态更新和阅读标记逻辑

+++ src/main/java/com/electricitybill/enums/BaseEnum.java
@@ -1,10 +1,10 @@
 package com.electricitybill.enums;
 
 public interface BaseEnum {
-    double getValue();
+    int getValue();
     String getDesc();
 
-    default boolean equalsValue(Double value){
+    default boolean equalsValue(Integer value){
         if (value == null) {
             return false;
         }
